+++ junifer/preprocess/warping/_fsl_warper.py
@@ -1,4 +1,4 @@
-"""Provide class for native space warping via ANTs antsApplyTransforms."""
+"""Provide class for space warping via FSL FLIRT."""
 
 # Authors: Synchon Mandal <s.mandal@fz-juelich.de>
 # License: AGPL
@@ -19,18 +19,18 @@ from ...pipeline import WorkDirManager
 from ...utils import logger, run_ext_cmd
 
 
-class ANTsNativeWarper:
-    """Class for native space warping via ANTs antsApplyTransforms.
+class FSLWarper:
+    """Class for space warping via FSL FLIRT.
 
-    This class uses ANTs' ``ResampleImage`` for resampling and
-    ``antsApplyTransforms`` for transformation.
+    This class uses FSL FLIRT's ``flirt`` for resampling and ``applywarp`` for
+    transformation.
 
     """
 
     _EXT_DEPENDENCIES: ClassVar[List[Dict[str, Union[str, List[str]]]]] = [
         {
-            "name": "ants",
-            "commands": ["ResampleImage", "antsApplyTransforms"],
+            "name": "fsl",
+            "commands": ["flirt", "applywarp"],
         },
     ]
 
@@ -41,7 +41,7 @@ class ANTsNativeWarper:
         input: Dict[str, Any],
         extra_input: Dict[str, Any],
     ) -> Dict[str, Any]:
-        """Preprocess using ANTs.
+        """Preprocess using FSL.
 
         Parameters
         ----------
@@ -59,7 +59,7 @@ class ANTsNativeWarper:
             reference file used for warping.
 
         """
-        logger.debug("Using ANTs for native space warping")
+        logger.debug("Using FSL for space warping")
 
         # Get the min of the voxel sizes from input and use it as the
         # resolution
@@ -67,47 +67,41 @@ class ANTsNativeWarper:
 
         # Create element-specific tempdir for storing post-warping assets
         element_tempdir = WorkDirManager().get_element_tempdir(
-            prefix="ants_warper"
+            prefix="fsl_warper"
         )
 
         # Create a tempfile for resampled reference output
-        resample_image_out_path = (
-            element_tempdir / "resampled_reference.nii.gz"
-        )
-        # Set ResampleImage command
-        resample_image_cmd = [
-            "ResampleImage",
-            "3",  # image dimension
-            f"{extra_input['T1w']['path'].resolve()}",
-            f"{resample_image_out_path.resolve()}",
-            f"{resolution}x{resolution}x{resolution}",
-            "0",  # option for spacing and not size
-            "3 3",  # Lanczos windowed sinc
+        flirt_out_path = element_tempdir / "resampled_reference.nii.gz"
+        # Set flirt command
+        flirt_cmd = [
+            "flirt",
+            "-interp spline",
+            f"-in {extra_input['T1w']['path'].resolve()}",
+            f"-ref {extra_input['T1w']['path'].resolve()}",
+            f"-applyisoxfm {resolution}",
+            f"-out {flirt_out_path.resolve()}",
         ]
-        # Call ResampleImage
-        run_ext_cmd(name="ResampleImage", cmd=resample_image_cmd)
+        # Call flirt
+        run_ext_cmd(name="flirt", cmd=flirt_cmd)
 
         # Create a tempfile for warped output
-        apply_transforms_out_path = element_tempdir / "output.nii.gz"
-        # Set antsApplyTransforms command
-        apply_transforms_cmd = [
-            "antsApplyTransforms",
-            "-d 3",
-            "-e 3",
-            "-n LanczosWindowedSinc",
+        applywarp_out_path = element_tempdir / "output.nii.gz"
+        # Set applywarp command
+        applywarp_cmd = [
+            "applywarp",
+            "--interp=spline",
             f"-i {input['path'].resolve()}",
-            # use resampled reference
-            f"-r {resample_image_out_path.resolve()}",
-            f"-t {extra_input['Warp']['path'].resolve()}",
-            f"-o {apply_transforms_out_path.resolve()}",
+            f"-r {flirt_out_path.resolve()}",  # use resampled reference
+            f"-w {extra_input['Warp']['path'].resolve()}",
+            f"-o {applywarp_out_path.resolve()}",
         ]
-        # Call antsApplyTransforms
-        run_ext_cmd(name="antsApplyTransforms", cmd=apply_transforms_cmd)
+        # Call applywarp
+        run_ext_cmd(name="applywarp", cmd=applywarp_cmd)
 
         # Load nifti
-        input["data"] = nib.load(apply_transforms_out_path)
+        input["data"] = nib.load(applywarp_out_path)
         # Save resampled reference path
-        input["reference_path"] = resample_image_out_path
+        input["reference_path"] = flirt_out_path
 
         # Use reference input's space as warped input's space
         input["space"] = extra_input["T1w"]["space"]
